requirements.txt
google-genai
pillow
python-dotenv
numpy
matplotlib
scikit-image
flask